requirements.txt
streamlit
numpy==1.23.5
opencv-python-headless
Pillow
tensorflow==2.12.0